+++ megatron/model/gpt2_model.py
@@ -81,7 +81,7 @@ class GPT2Model(MegatronModule):
             return output
         else:
             if self.fp16_lm_cross_entropy:
-                assert output.dtype == torch.half
+                assert output.dtype == torch.float
                 loss = mpu.vocab_parallel_cross_entropy(output, labels)
             else:
                 loss = mpu.vocab_parallel_cross_entropy(output.float(), labels)
